Fix price < 0

+++ menus/models.py
@@ -10,7 +10,7 @@ class Restaurant(models.Model):
 class MenuItem(models.Model):
     restaurant = models.ForeignKey(Restaurant, on_delete=models.CASCADE)
     name = models.CharField(max_length=100)
-    price = models.DecimalField(max_digits=8, decimal_places=2)
+    price = models.PositiveIntegerField()
 
     def __str__(self):
         return self.name
